requirements.txt
streamlit
groq
chromadb
semantic-router
pandas
python-dotenv
sqlite3
pathlib
